+++ include/ui/setup.hpp
@@ -290,7 +290,8 @@ namespace nwo5::ui {
             return {m_node};
         }
 
-        Setup id(const std::string& pID) requires std::derived_from<Node, cocos2d::CCNode> {
+        template<typename T>
+        Setup id(T pID) requires std::derived_from<Node, cocos2d::CCNode> {
             m_node->setID(pID);
             return {m_node};
         }