+++ changelog.md
@@ -5,6 +5,7 @@
 - make ui header include some of geodes ui stuff
 - events for saved settings
 - little changes to readme
+- make hook prio function a bit more clear
 
 ## v1.8.0
 **Added**